Dialog system functionality update
Added two more functions to dialog system. A time out function, allowing you to close the text box after a certain time and abilty to sign the text box with name of the character with function OpenDialogWithName, so the text box displays a name of the person that is saying given line.